finacebot/requirements.txt
python-telegram-bot==20.5
pymysql==1.1.0
aiomysql==0.1.0
aiohttp==3.8.4
python-dotenv==1.0.1
pandas==2.2.3
matplotlib==3.8.2
